social number updated to longlong

+++ GroupFour_DoctorsOffice/patient.cpp
@@ -1,6 +1,6 @@
 #include "patient.h"
 
-Patient::Patient(int social_number, QString first_name, QString last_name, int phone_number, int doctor_id) : Person(first_name, last_name, phone_number)
+Patient::Patient(long long social_number, QString first_name, QString last_name, int phone_number, int doctor_id) : Person(first_name, last_name, phone_number)
 {
     this->social_number_ = social_number;
     this->first_name_ = first_name;
@@ -14,7 +14,7 @@ int Patient::getDoctorID()
     return this->doctor_id_;
 }
 
-int Patient::getSocialNumber()
+long long Patient::getSocialNumber()
 {
     return this->social_number_;
 }
